Исправлены неточности в REST
