+++ javadoc/animal/shelter/animalsshelter/controllers/web/AddController.html
@@ -124,7 +124,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>animal.shelter.animalsshelter.controllers.web.AddController</li>
+<li>animal.shelter.animalsshelter.controllers.web.DogController</li>
 </ul>
 </li>
 </ul>
