+++ javadoc/animal/shelter/animalsshelter/controllers/web/GetController.html
@@ -124,7 +124,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>animal.shelter.animalsshelter.controllers.web.GetController</li>
+<li>animal.shelter.animalsshelter.controllers.web.UserController</li>
 </ul>
 </li>
 </ul>
