+++ src/main/java/animal/shelter/animalsshelter/controllers/stateTest/BotContext.java
@@ -6,7 +6,6 @@ import lombok.Data;
 
 @Data
 public class BotContext {
-
     private final TelegramBotStart bot;
     private final TestUser user;
     private final String input;
