+++ src/main/java/animal/shelter/animalsshelter/controllers/web/ReportController.java
@@ -1,0 +1,35 @@
+package animal.shelter.animalsshelter.controllers.web;
+
+import animal.shelter.animalsshelter.model.Report;
+import animal.shelter.animalsshelter.service.ReportService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import java.util.List;
+
+@RestController
+@RequestMapping("/report")
+public class ReportController {
+
+    private final ReportService reportService;
+
+    public ReportController(ReportService reportService) {
+        this.reportService = reportService;
+    }
+
+    @GetMapping("/all")
+    @Operation(summary = "Получить список отправленных отчётов",
+            description = "Данный запрос позволяет получить полный список отправленных отчётов когда-либо")
+    @ApiResponse(responseCode = "200",
+            description = "Операция успешна")
+    @ApiResponse(responseCode = "400",
+            description = "параметры запроса отсутствуют или имеют некорректный формат;")
+    @ApiResponse(responseCode = "500",
+            description = "произошла ошибка, не зависящая от вызывающей стороны.")
+    public List<Report> getAllReports() {
+        List<Report> reports = reportService.getAllReports();
+        return reports;
+    }
+}
